backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==4.2.6
django-colorfield==0.10.1
django-cors-headers==4.3.0
django-filter==23.3
django-import-export==3.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
et-xmlfile==1.1.0
filetype==1.2.0
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
idna==3.4
isort==5.12.0
kombu==5.3.2
MarkupPy==1.14
mccabe==0.7.0
mypy==1.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pep8==1.7.1
pep8-naming==0.13.3
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.9
webcolors==1.13
xlrd==2.0.1
xlwt==1.3.0
